cereate comment area
