+++ public/like.tsx
@@ -9,7 +9,7 @@ export const Like = () => {
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
-      <g clip-path="url(#clip0_423_436)">
+      <g clipPath="url(#clip0_423_436)">
         <mask
           id="mask0_423_436"
           maskUnits="userSpaceOnUse"
