+++ public/liked.tsx
@@ -9,7 +9,7 @@ export const Liked = () => {
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
-      <g clip-path="url(#clip0_1820_1020)">
+      <g clipPath="url(#clip0_1820_1020)">
         <mask
           id="mask0_1820_1020"
           maskUnits="userSpaceOnUse"
