+++ src/app/player/page.tsx
@@ -1,107 +1,47 @@
 'use client';
 import * as React from 'react';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-import Slider from '@mui/material/Slider';
-import { PlayButton } from '@public/playButton';
-import { StopButton } from '@public/stopButton';
+import { Search } from '@/components/ui/Search';
+import { Comment } from '@/components/ui/Comment';
+import { MusicPlayer } from '@/components/ui/MusicPlayer';
 import { useState } from 'react';
-import { useEffect } from 'react';
-import { Tag } from '@/components/ui/Tag';
-import { Liked } from '@public/liked';
-import { Like } from '@public/like';
-import profileImage from '../../../public/profImage.png';
 
 export default function ContinuousSlider() {
-  const theme = createTheme({
-    components: {
-      MuiSlider: {
-        styleOverrides: {
-          track: {
-            '&.MuiSlider-track': {
-              backgroundColor: '#565656', // トラックの色
-              border: 'none',
-              opacity: 1,
-            },
-          },
-          thumb: {
-            color: '#565656', // スライダーのつまみの色
-          },
-          rail: {
-            '&.MuiSlider-rail': {
-              backgroundColor: '#565656', // レール（トラックの裏側）の色
-              opacity: 1, // レール（トラックの裏側）の色
-            },
-          },
-        },
-      },
-    },
-  });
-
-  const [isPlay, setIsPlay] = useState<boolean>(false);
-
-  const onClickFunc = () => {
-    setIsPlay(!isPlay);
-  };
-  useEffect(() => {
-    console.log(isPlay);
-  }, [isPlay]);
+  const [isClick, setIsClick] = useState<boolean>(false);
+  React.useEffect(() => {
+    console.log(isClick);
+  }, [isClick]);
   return (
-    <div className="w-11/12 rounded-xl bg-[#E3DEDA]">
-      <div className="flex flex-col items-center">
-        <div className="relative h-10 w-fit flex flex-col items-center mt-8">
-          <p className="text-2xl font-bold text-[#646767] z-20 relative">
-            夜もすがら君を想う
-          </p>
-          <label className="w-full h-7 bg-[#B3D0CF] block absolute top-3 rounded-xl"></label>
+    <>
+      <div className="w-11/12 mx-auto">
+        <div className="my-6 ml-auto w-fit">
+          <Search />
         </div>
-        <div className="mt-6">
-          {isPlay ? (
-            <button
-              onClick={() => {
-                onClickFunc();
-              }}
-            >
-              <PlayButton />
-            </button>
-          ) : (
-            <button
-              onClick={() => {
-                onClickFunc();
-              }}
-            >
-              <StopButton />
-            </button>
-          )}
-        </div>
-        <div className="w-3/4">
-          <ThemeProvider theme={theme}>
-            <Slider disabled defaultValue={30} aria-label="Disabled slider" />
-          </ThemeProvider>
-        </div>
-        <div className="w-5/6">
-          <div className="flex w-full justify-between mb-3">
-            <Tag text="tag" href="/player" />
-            <button className="">
-              <Like />
-            </button>
-          </div>
-          <p className="text-xs text-[#646767] mb-6">
-            概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要概要
-          </p>
-          <div className="flex justify-between items-center mb-4">
-            <div className="flex items-center">
-              <img
-                src={profileImage.src}
-                className="block w-11 h-11 rounded-full"
-              />
-              <p className="text-[#646767] font-bold ml-3">mocha</p>
-            </div>
-            <button className="w-16 h-8 bg-[#646767] text-[#DDBFAE] rounded-md">
-              follow
-            </button>
-          </div>
-        </div>
+        <MusicPlayer />
+        <Comment isClick={isClick} setIsClick={setIsClick} />
       </div>
-    </div>
+      {isClick ? (
+        <div className="flex items-center justify-center">
+          <label
+            className="w-[100vh] h-[100vh] bg-black opacity-20 top-0 fixed z-20"
+            onClick={() => {
+              setIsClick(false);
+            }}
+          ></label>
+          <div className="flex flex-col bg-white w-2/3 h-80 z-30 fixed top-44 items-center rounded-xl">
+            <div className="w-4/5 flex flex-col items-end">
+              <button className="w-16 h-8 bg-[#646767] rounded-md text-[#DDBFAE] mt-7 mb-6">
+                send
+              </button>
+              <textarea
+                placeholder="コメントを書き込む"
+                className="w-full h-44 rounded-lg border border-[#6E96A5] py-3 px-2 placeholder:text-xs"
+              />
+            </div>
+          </div>
+        </div>
+      ) : (
+        <div></div>
+      )}
+    </>
   );
 }
